requirements.txt
beautifulsoup4==4.12.3
capsolver==1.0.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
fuzzywuzzy==0.18.0
idna==3.10
nextcaptcha-python==1.1.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
